+++ custom_components/fotmob_leagues/const.py
@@ -5,5 +5,7 @@ DOMAIN = "fotmob_leagues"
 CONF_LEAGUE_ID = "league_id"
 
 API_URL = "https://www.fotmob.com/api/data/leagues"
-DEEP_STATS_API_URL = "https://www.fotmob.com/api/data/leagueseasondeepstats"
+LEAGUE_LOGO_URL = (
+    "https://images.fotmob.com/image_resources/logo/leaguelogo/{league_id}.png"
+)
 UPDATE_INTERVAL_MINUTES = 30